refactor: remove sortable payment methods list (#1792)
