--- client/components/orderable-list/style.scss
@@ -1,38 +1,0 @@
-.orderable-list {
-	margin: 0;
-
-	&__item {
-		display: flex;
-		margin: 0;
-		padding: 16px 24px 14px 24px;
-		box-shadow: inset 0 -1px 0 #e8eaeb;
-		background: #fff;
-
-		@include breakpoint( '>660px' ) {
-			padding: 24px 33px 24px 24px;
-		}
-
-		&.is-dragged {
-			box-shadow: none;
-			border-bottom: none;
-		}
-	}
-
-	&__drag-handle-wrapper {
-		flex: 0 0 28px;
-		margin-right: 16px;
-	}
-
-	&__drag-handle {
-		&#{&} {
-			display: none;
-		}
-
-		.has-drag-handles & {
-			display: inline-flex;
-			background: url( '../../../assets/images/drag-handle.svg' )
-				no-repeat center center;
-			background-size: 10px 6px;
-		}
-	}
-}
