+++ client/components/payment-methods-list/delete-button.js
