+++ client/components/payment-methods-list/delete-button.scss
