+++ client/components/payment-methods-list/payment-method.js
@@ -2,13 +2,12 @@
 /**
  * External dependencies
  */
-import { __ } from '@wordpress/i18n';
+import classNames from 'classnames';
 
 /**
  * Internal dependencies
  */
 import './payment-method.scss';
-import { getPaymentMethodSettingsUrl } from '../../utils';
 import PaymentMethodDeleteButton from './delete-button';
 
 const PaymentMethod = ( {
@@ -17,29 +16,26 @@ const PaymentMethod = ( {
 	Icon = () => null,
 	description,
 	onDeleteClick,
+	className,
 } ) => {
-	const settingsUrl = getPaymentMethodSettingsUrl( id );
-
 	return (
-		<>
+		<li
+			className={ classNames(
+				'payment-method',
+				{ 'has-icon-border': 'cc' !== id },
+				className
+			) }
+		>
 			<div className="payment-method__icon">
 				<Icon />
 			</div>
 			<div className="payment-method__text">
-				<a href={ settingsUrl } className="payment-method__label">
-					{ label }
-				</a>
+				<div className="payment-method__label">{ label }</div>
 				<div className="payment-method__description">
 					{ description }
 				</div>
 			</div>
 			<div className="payment-method__actions">
-				<a
-					href={ settingsUrl }
-					className="payment-method__action manage"
-				>
-					{ __( 'Manage', 'woocommerce-payments' ) }
-				</a>
 				{ onDeleteClick && (
 					<PaymentMethodDeleteButton
 						className="payment-method__action delete"
@@ -50,7 +46,7 @@ const PaymentMethod = ( {
 					/>
 				) }
 			</div>
-		</>
+		</li>
 	);
 };
