+++ client/components/payment-methods-list/style.scss
@@ -1,0 +1,3 @@
+.payment-methods-list {
+	margin: 0;
+}
