+++ client/components/payment-methods-list/test/payment-method.js
@@ -19,45 +19,14 @@ describe( 'PaymentMethod', () => {
 		expect( screen.queryByText( 'Bar' ) ).toBeInTheDocument();
 	} );
 
-	test( 'renders "Manage" and "Delete"', () => {
-		render( <PaymentMethod label="Foo" onDeleteClick={ jest.fn() } /> );
-
-		const manageLink = screen.queryByRole( 'link', {
-			name: 'Manage',
-		} );
-
-		const deleteButton = screen.queryByRole( 'button', {
-			name: 'Delete',
-		} );
-
-		expect( manageLink ).toBeInTheDocument();
-		expect( deleteButton ).toBeInTheDocument();
-	} );
-
 	test( 'does not render "Delete" when the handler is not provided', () => {
 		render( <PaymentMethod label="Foo" onDeleteClick={ undefined } /> );
 
-		const manageLink = screen.queryByRole( 'link', {
-			name: 'Manage',
-		} );
-
 		const deleteButton = screen.queryByRole( 'button', {
 			name: 'Delete',
 		} );
 
-		expect( manageLink ).toBeInTheDocument();
 		expect( deleteButton ).not.toBeInTheDocument();
-	} );
-
-	test( '"Manage" is a link to expected URL', () => {
-		render( <PaymentMethod id="foo" label="Bar" /> );
-
-		const manageLink = screen.getByRole( 'link', {
-			name: 'Manage',
-		} );
-		expect( manageLink.getAttribute( 'href' ) ).toEqual(
-			'admin.php?page=wc-settings&tab=checkout&section=woocommerce_payments_foo'
-		);
 	} );
 
 	test( 'clicking the "Delete" button with confirmation calls onDeleteClick()', () => {
@@ -96,16 +65,5 @@ describe( 'PaymentMethod', () => {
 
 		expect( handleDeleteClickMock ).toHaveBeenCalledTimes( 1 );
 		expect( handleDeleteClickMock ).toHaveBeenCalledWith( 'foo' );
-	} );
-
-	test( 'label is a link to expected URL', () => {
-		render( <PaymentMethod id="foo" label="Bar" /> );
-
-		const paymentMethodLabel = screen.getByRole( 'link', {
-			name: 'Bar',
-		} );
-		expect( paymentMethodLabel.getAttribute( 'href' ) ).toEqual(
-			'admin.php?page=wc-settings&tab=checkout&section=woocommerce_payments_foo'
-		);
 	} );
 } );
