+++ client/payment-methods/index.js
@@ -4,7 +4,7 @@
  */
 import React from 'react';
 import { __ } from '@wordpress/i18n';
-import { Card, CardBody } from '@wordpress/components';
+import { Card, CardBody, CardDivider } from '@wordpress/components';
 import classNames from 'classnames';
 
 /**
@@ -12,8 +12,8 @@ import classNames from 'classnames';
  */
 import './style.scss';
 import { useEnabledPaymentMethodIds } from 'data';
-import OrderableList from 'components/orderable-list';
-import PaymentMethod from 'components/orderable-list/payment-method';
+import PaymentMethodsList from 'components/payment-methods-list';
+import PaymentMethod from 'components/payment-methods-list/payment-method';
 import PaymentMethodsSelector from 'settings/payment-methods-selector';
 import CreditCardIcon from '../gateway-icons/credit-card';
 import GiropayIcon from '../gateway-icons/giropay';
@@ -79,42 +79,16 @@ const PaymentMethods = () => {
 		);
 	};
 
-	const handleDragEnd = ( event ) => {
-		const { active, over } = event;
-
-		if ( active.id !== over.id ) {
-			const oldIndex = enabledMethodIds.indexOf( active.id );
-			const newIndex = enabledMethodIds.indexOf( over.id );
-
-			const enabledMethodIdsCopy = [ ...enabledMethodIds ];
-			enabledMethodIdsCopy.splice(
-				0 > newIndex
-					? enabledMethodIdsCopy.length + newIndex
-					: newIndex,
-				0,
-				enabledMethodIdsCopy.splice( oldIndex, 1 )[ 0 ]
-			);
-
-			updateEnabledMethodIds( enabledMethodIdsCopy );
-		}
-	};
-
 	return (
 		<>
 			<Card className="payment-methods">
-				<CardBody className="payment-methods__enabled-methods-container">
-					<OrderableList
-						onDragEnd={ handleDragEnd }
-						className="payment-methods__enabled-methods"
-					>
+				<CardBody size={ null }>
+					<PaymentMethodsList className="payment-methods__enabled-methods">
 						{ enabledMethods.map(
 							( { id, label, description, Icon } ) => (
 								<PaymentMethod
 									key={ id }
 									Icon={ Icon }
-									className={ classNames( 'payment-method', {
-										'has-icon-border': 'cc' !== id,
-									} ) }
 									onDeleteClick={
 										1 < enabledMethods.length
 											? handleDeleteClick
@@ -126,8 +100,9 @@ const PaymentMethods = () => {
 								/>
 							)
 						) }
-					</OrderableList>
+					</PaymentMethodsList>
 				</CardBody>
+				<CardDivider />
 				<CardBody className="payment-methods__available-methods-container">
 					<PaymentMethodsSelector className="payment-methods__add-payment-method" />
 					<ul className="payment-methods__available-methods">
