+++ client/payment-methods/style.scss
@@ -1,23 +1,7 @@
 .payment-methods {
 	&__available-methods-container {
+		@include modal-footer-buttons-left;
 		align-items: center;
-		display: flex;
-
-		&#{&} {
-			&#{&} {
-				padding: 24px;
-
-				@include breakpoint( '>660px' ) {
-					padding: 24px 32px;
-				}
-			}
-		}
-	}
-
-	&__add-payment-method {
-		&#{&} {
-			margin-right: 16px;
-		}
 	}
 
 	&__available-methods {
@@ -41,14 +25,6 @@
 
 		&:last-child {
 			margin-right: 1px; // 1px to accommodate for box-shadow
-		}
-	}
-
-	&__enabled-methods-container {
-		&#{&} {
-			&#{&} {
-				padding: 0;
-			}
 		}
 	}
 }
