+++ client/payment-methods/test/index.js
@@ -75,7 +75,7 @@ describe( 'PaymentMethods', () => {
 		} );
 	} );
 
-	test( 'enabled methods are rendered with "Manage" and "Delete" buttons', () => {
+	test( 'enabled methods are rendered with "Delete" buttons', () => {
 		useEnabledPaymentMethodIds.mockReturnValue( {
 			enabledPaymentMethodIds: [
 				'woocommerce_payments',
@@ -88,9 +88,6 @@ describe( 'PaymentMethods', () => {
 		const cc = screen.getByText( 'Credit card / debit card' );
 		const listItem = cc.closest( 'li' );
 
-		expect(
-			within( listItem ).queryByRole( 'link', { name: 'Manage' } )
-		).toBeInTheDocument();
 		expect(
 			within( listItem ).queryByRole( 'button', { name: 'Delete' } )
 		).toBeInTheDocument();
@@ -106,9 +103,6 @@ describe( 'PaymentMethods', () => {
 		const cc = screen.getByText( 'Credit card / debit card' );
 		const listItem = cc.closest( 'li' );
 
-		expect(
-			within( listItem ).queryByRole( 'link', { name: 'Manage' } )
-		).toBeInTheDocument();
 		expect(
 			within( listItem ).queryByRole( 'button', { name: 'Delete' } )
 		).not.toBeInTheDocument();
