+++ client/stylesheets/abstracts/_mixins.scss
@@ -30,3 +30,13 @@
 		}
 	}
 }
+
+@mixin modal-footer-buttons-left {
+	display: flex;
+
+	> * {
+		&:not( :last-child ) {
+			margin-right: 16px;
+		}
+	}
+}
